Update customer portal handling command
